Propagate ChannelHash u64 widening into Node / Python bindings + Go header
The earlier substrate change (ChannelHash from u32 to u64) widened
PermissionToken::channel_hash and the FFI's net_channel_hash
out-pointer type. The Node / Python NAPI / PyO3 bindings and the
Go-binding header weren't updated in lockstep and now fail to
build.

bindings/node/src/identity.rs:
  - TokenInfo::channel_hash becomes BigInt (u64 from a JS number
    would silently truncate at 2^53).
  - parse_token wraps the field with BigInt::from(...).
  - pub fn channel_hash(channel) returns Result<BigInt> instead
    of Result<u32>, mirroring the not_before / not_after BigInt
    convention.

bindings/python/src/identity.rs:
  - pub fn channel_hash returns PyResult<u64>. Python's int is
    arbitrary-precision so the u64 round-trips without truncation.
    The internal pyo3 dict-set path auto-converts.

include/net.go.h:
  - net_channel_hash's out-parameter typedef switches from
    uint32_t* to uint64_t*, matching the Rust-side signature
    bumped in the prior commit. Comment updated.

cargo check --workspace --all-features now passes; no other
binding surfaces reference the canonical channel_hash width
directly.

+++ net/crates/net/bindings/python/src/identity.rs
@@ -341,10 +341,12 @@ pub fn delegate_token(
     Ok(child.to_bytes())
 }
 
-/// Hash a channel name to its canonical 32-bit substrate identifier
+/// Hash a channel name to its canonical 64-bit substrate identifier
 /// (used for ACL / config / storage keys; the wire `NetHeader`
-/// fast-path hint is the low 16 bits of this value).
+/// fast-path hint is the low 16 bits of this value). Python's
+/// `int` is arbitrary-precision so the full u64 round-trips
+/// without truncation.
 #[pyfunction]
-pub fn channel_hash(channel: &str) -> PyResult<u32> {
+pub fn channel_hash(channel: &str) -> PyResult<u64> {
     channel_to_hash(channel)
 }
